Add working Channel 0
Channel 0 is responsible for the "behind the scenes" communication with RabbitMQ.

It is where the connection is negotiated, heartbeats are handled, connection.blocked and unblocked notifications are sent etc.

As of this commit, this will negotiate a connection with RabbitMQ and handle all the various things that need to be handled.