Update the last heartbeat timeout when frames are received #7
I don't think it's the root of the issue, but it will help.

"Any traffic (e.g. protocol operations, published messages, acknowledgements) counts for a valid heartbeat. Clients may choose to send heartbeat frames regardless of whether there was any other traffic on the connection but some only do it when necessary."